refactor(src/utils/extensions.ts):  externalise Quick Pick UI tools
