+++ src/ui/extensionsQuickPick.ts
@@ -1,0 +1,54 @@
+import * as vscode from "vscode";
+import { getGitHubLink, formatExtensionLabel } from "../utils/extensions";
+
+export interface ExtensionQuickPickItem extends vscode.QuickPickItem {
+	url?: string;
+}
+
+export function createExtensionItems(extensions: string[]): ExtensionQuickPickItem[] {
+	return extensions.map((ext) => ({
+		label: formatExtensionLabel(ext),
+		description: ext,
+		buttons: [
+			{
+				iconPath: new vscode.ThemeIcon("github"),
+				tooltip: "Open GitHub Repository",
+			},
+		],
+		url: getGitHubLink(ext),
+	}));
+}
+
+export async function showExtensionQuickPick(
+	extensionsList: string[],
+	recentlyInstalled: string[]
+): Promise<readonly ExtensionQuickPickItem[]> {
+	const groupedExtensions: ExtensionQuickPickItem[] = [
+		{
+			label: "Recently Installed",
+			kind: vscode.QuickPickItemKind.Separator,
+		},
+		...createExtensionItems(recentlyInstalled),
+		{
+			label: "All Extensions",
+			kind: vscode.QuickPickItemKind.Separator,
+		},
+		...createExtensionItems(extensionsList.filter((ext) => !recentlyInstalled.includes(ext))).sort((a, b) =>
+			a.label.localeCompare(b.label)
+		),
+	];
+
+	const quickPick = vscode.window.createQuickPick<ExtensionQuickPickItem>();
+	quickPick.items = groupedExtensions;
+	quickPick.placeholder = "Select Quarto extensions to install";
+	quickPick.canSelectMany = true;
+	quickPick.matchOnDescription = true;
+
+	return new Promise((resolve) => {
+		quickPick.onDidAccept(() => {
+			resolve(quickPick.selectedItems);
+			quickPick.hide();
+		});
+		quickPick.show();
+	});
+}
